quick_protobuf_codec

Struct Codec

pub struct Codec<In, Out = In> { /* private fields */ }

Codec implements Encoder and Decoder, uses unsigned_varint to prefix messages with their length and uses quick_protobuf and a provided struct implementing MessageRead and MessageWrite to do the encoding.

Implementations

impl<In, Out> Codec<In, Out>

pub fn new(max_message_len_bytes: usize) -> Self

Create new Codec.

Parameter max_message_len_bytes determines the maximum length of the Protobuf message. The limit does not include the bytes needed for the unsigned_varint.